Add files via upload
Problem : A person wants to transfer bananas over to a destination A km away. He initially has
B bananas and a camel. The camel cannot carry more than C bananas at a time and eats a
banana every km it travels. Given three integers A, B, and C, the task is to find the maximum
number of bananas the person can transfer to the destination using the camel.